+++ src/singletons/Settings.hpp
@@ -6,9 +6,11 @@
 #include "BaseSettings.hpp"
 #include "common/Channel.hpp"
 #include "common/SignalVector.hpp"
+#include "controllers/filters/FilterRecord.hpp"
 #include "controllers/highlights/HighlightPhrase.hpp"
 #include "controllers/moderationactions/ModerationAction.hpp"
 #include "singletons/Toasts.hpp"
+#include "util/StreamerMode.hpp"
 #include "widgets/Notebook.hpp"
 
 using TimeoutButton = std::pair<QString, int>;
@@ -19,6 +21,7 @@ class HighlightPhrase;
 class HighlightBlacklistUser;
 class IgnorePhrase;
 class TaggedUser;
+class FilterRecord;
 
 /// Settings which are availlable for reading on all threads.
 class ConcurrentSettings
@@ -31,6 +34,7 @@ public:
     SignalVector<HighlightBlacklistUser> &blacklistedUsers;
     SignalVector<IgnorePhrase> &ignoredMessages;
     SignalVector<QString> &mutedChannels;
+    SignalVector<FilterRecordPtr> &filterRecords;
     //SignalVector<TaggedUser> &taggedUsers;
     SignalVector<ModerationAction> &moderationActions;
 
@@ -177,7 +181,18 @@ public:
     BoolSetting unshortLinks = {"/links/unshortLinks", false};
     BoolSetting lowercaseDomains = {"/links/linkLowercase", true};
 
-    /// Ignored phrases
+    /// Streamer Mode
+    EnumSetting<StreamerModeSetting> enableStreamerMode = {
+        "/streamerMode/enabled", StreamerModeSetting::DetectObs};
+    BoolSetting streamerModeHideUsercardAvatars = {
+        "/streamerMode/hideUsercardAvatars", true};
+    BoolSetting streamerModeHideLinkThumbnails = {
+        "/streamerMode/hideLinkThumbnails", true};
+    BoolSetting streamerModeHideViewerCountAndDuration = {
+        "/streamerMode/hideViewerCountAndDuration", false};
+    BoolSetting streamerModeMuteMentions = {"/streamerMode/muteMentions", true};
+
+    /// Ignored Phrases
     QStringSetting ignoredPhraseReplace = {"/ignore/ignoredPhraseReplace",
                                            "***"};
 
@@ -244,6 +259,10 @@ public:
     QStringSetting highlightColor = {"/highlighting/color", ""};
 
     BoolSetting longAlerts = {"/highlighting/alerts", false};
+
+    /// Filtering
+    BoolSetting excludeUserMessagesFromFilter = {
+        "/filtering/excludeUserMessages", false};
 
     /// Logging
     BoolSetting enableLogging = {"/logging/enabled", false};
@@ -316,8 +335,6 @@ public:
     BoolSetting restartOnCrash = {"/misc/restartOnCrash", false};
     BoolSetting attachExtensionToAnyProcess = {
         "/misc/attachExtensionToAnyProcess", false};
-    BoolSetting hideViewerCountAndDuration = {
-        "/misc/hideViewerCountAndDuration", false};
     BoolSetting askOnImageUpload = {"/misc/askOnImageUpload", true};
 
     /// Debug