+++ src/widgets/splits/SplitHeader.cpp
@@ -12,6 +12,7 @@
 #include "singletons/WindowManager.hpp"
 #include "util/LayoutCreator.hpp"
 #include "util/LayoutHelper.hpp"
+#include "util/StreamerMode.hpp"
 #include "widgets/Label.hpp"
 #include "widgets/TooltipWidget.hpp"
 #include "widgets/dialogs/SettingsDialog.hpp"
@@ -86,30 +87,64 @@ namespace {
     }
     auto formatTooltip(const TwitchChannel::StreamStatus &s, QString thumbnail)
     {
-        return QString("<style>.center { text-align: center; }</style> \
-            <p class=\"center\">%1%2%3%4%5%6 for %7 with %8 viewers</p>")
-            .arg(s.title.toHtmlEscaped())
-            .arg(s.title.isEmpty() ? QString() : "<br><br>")
-            .arg(getSettings()->thumbnailSizeStream.getValue() > 0
-                     ? ((thumbnail.isEmpty()
-                             ? "Couldn't fetch thumbnail"
-                             : "<img src=\"data:image/jpg;base64, " +
-                                   thumbnail + "\"/>") +
-                        "<br>")
-                     : QString())
-            .arg(s.game.toHtmlEscaped())
-            .arg(s.game.isEmpty() ? QString() : "<br>")
-            .arg(s.rerun ? "Vod-casting" : "Live")
-            .arg(getSettings()->hideViewerCountAndDuration ? "&lt;Hidden&gt;"
-                                                           : s.uptime)
-            .arg(getSettings()->hideViewerCountAndDuration
-                     ? "&lt;Hidden&gt;"
-                     : QString::number(s.viewerCount));
+        auto title = [&s]() -> QString {
+            if (s.title.isEmpty())
+            {
+                return QStringLiteral("");
+            }
+
+            return s.title.toHtmlEscaped() + "<br><br>";
+        }();
+
+        auto tooltip = [&thumbnail]() -> QString {
+            if (getSettings()->thumbnailSizeStream.getValue() == 0)
+            {
+                return QStringLiteral("");
+            }
+
+            if (thumbnail.isEmpty())
+            {
+                return QStringLiteral("Couldn't fetch thumbnail<br>");
+            }
+
+            return "<img src=\"data:image/jpg;base64, " + thumbnail + "\"><br>";
+        }();
+
+        auto game = [&s]() -> QString {
+            if (s.game.isEmpty())
+            {
+                return QStringLiteral("");
+            }
+
+            return s.game.toHtmlEscaped() + "<br>";
+        }();
+
+        auto extraStreamData = [&s]() -> QString {
+            if (isInStreamerMode() &&
+                getSettings()->streamerModeHideViewerCountAndDuration)
+            {
+                return QStringLiteral(
+                    "<span style=\"color: #808892;\">&lt;Streamer "
+                    "Mode&gt;</span>");
+            }
+
+            return QString("%1 for %2 with %3 viewers")
+                .arg(s.rerun ? "Vod-casting" : "Live")
+                .arg(s.uptime)
+                .arg(QString::number(s.viewerCount));
+        }();
+
+        return QString("<p style=\"text-align: center;\">" +  //
+                       title +                                //
+                       tooltip +                              //
+                       game +                                 //
+                       extraStreamData +                      //
+                       "</p>"                                 //
+        );
     }
     auto formatOfflineTooltip(const TwitchChannel::StreamStatus &s)
     {
-        return QString("<style>.center { text-align: center; }</style> \
-                       <p class=\"center\">Offline<br>%1</p>")
+        return QString("<p style=\"text-align: center;\">Offline<br>%1</p>")
             .arg(s.title.toHtmlEscaped());
     }
     auto formatTitle(const TwitchChannel::StreamStatus &s, Settings &settings)
@@ -228,6 +263,11 @@ void SplitHeader::initializeLayout()
                     }
                 });
         }),
+        // viewer list
+        this->viewersButton_ = makeWidget<Button>([&](auto w) {
+            QObject::connect(w, &Button::leftClicked, this,
+                             [this]() { this->split_->showViewerList(); });
+        }),
         // dropdown
         this->dropdownButton_ = makeWidget<Button>([&](auto w) {
             /// XXX: this never gets disconnected
@@ -287,6 +327,7 @@ std::unique_ptr<QMenu> SplitHeader::createMainMenu()
                     QKeySequence("Ctrl+N"));
     menu->addAction("Search", this->split_, &Split::showSearch,
                     QKeySequence("Ctrl+F"));
+    menu->addAction("Set filters", this->split_, &Split::setFiltersDialog);
     menu->addSeparator();
 #ifdef USEWEBENGINE
     this->dropdownMenu.addAction("Start watching", this, [this] {
@@ -547,16 +588,27 @@ void SplitHeader::initializeModeSignals(EffectLabel &label)
     });
 }
 
+void SplitHeader::resetThumbnail()
+{
+    this->lastThumbnail_.invalidate();
+    this->thumbnail_.clear();
+}
+
 void SplitHeader::handleChannelChanged()
 {
+    this->resetThumbnail();
+
+    this->updateChannelText();
+
     this->channelConnections_.clear();
 
     auto channel = this->split_->getChannel();
     if (auto twitchChannel = dynamic_cast<TwitchChannel *>(channel.get()))
     {
         this->channelConnections_.emplace_back(
-            twitchChannel->liveStatusChanged.connect(
-                [this]() { this->updateChannelText(); }));
+            twitchChannel->liveStatusChanged.connect([this]() {
+                this->updateChannelText();  //
+            }));
     }
 }
 
@@ -567,12 +619,18 @@ void SplitHeader::scaleChangedEvent(float scale)
     this->setFixedHeight(w);
     this->dropdownButton_->setFixedWidth(w);
     this->moderationButton_->setFixedWidth(w);
+    this->viewersButton_->setFixedWidth(w);
     this->addButton_->setFixedWidth(w * 5 / 8);
 }
 
 void SplitHeader::setAddButtonVisible(bool value)
 {
     this->addButton_->setVisible(value);
+}
+
+void SplitHeader::setViewersButtonVisible(bool value)
+{
+    this->viewersButton_->setVisible(value);
 }
 
 void SplitHeader::updateChannelText()
@@ -641,6 +699,11 @@ void SplitHeader::updateChannelText()
         {
             this->tooltipText_ = formatOfflineTooltip(*streamStatus);
         }
+    }
+
+    if (!title.isEmpty() && this->split_->getFilters().size() != 0)
+    {
+        title += " - filtered";
     }
 
     this->titleLabel_->setText(title.isEmpty() ? "<empty>" : title);
@@ -780,11 +843,13 @@ void SplitHeader::themeChangedEvent()
     // --
     if (this->theme->isLightTheme())
     {
+        this->viewersButton_->setPixmap(getResources().buttons.viewersDark);
         this->dropdownButton_->setPixmap(getResources().buttons.menuDark);
         this->addButton_->setPixmap(getResources().buttons.addSplit);
     }
     else
     {
+        this->viewersButton_->setPixmap(getResources().buttons.viewersLight);
         this->dropdownButton_->setPixmap(getResources().buttons.menuLight);
         this->addButton_->setPixmap(getResources().buttons.addSplitDark);
     }